Added feature to alter the appearance of the log line called by journalctl.
The relevant option is the '-o' flag which can be called with a formatter
string, e.g.

    journalctl -o "format:{__REALTIME_TIMESTAMP:date} {MESSAGE:60}"

which will output the date (a formatted Unix timestamp) and the message
(truncated to 60 characters). Literals are also allowed outside of the curly
braces. There are only the mentioned two options - date formatting and
string truncating - available. Those cannot be mixed.

The possible date formats are

    - date              Okt 20 10:22:05
    - date-iso          2015-10-20T10:22:05+0200
    - date-precise      Okt 20 10:22:05.056146
    - date-raw          [1445329325.056146]

where 'date' is the default on "__REALTIME_TIMESTAMP".